# docs/openapi/creators-api.yaml
...
openapi: 3.1.0
info:
  title: qoe.fi — API Créateurs & Médias
  version: 1.0.0
  description: >-
    API pour les médias et créateurs qui utilisent qoe.fi comme plateforme de
    diffusion (publier sur qoe.fi depuis leur CMS) ou comme CMS headless
    (diffuser leurs articles sur leur propre site). Contrat stable — les
    changements sont versionnés, jamais cassants.
  contact:
    name: qoe.fi
    url: https://qoe.fi
servers:
  - url: https://api.qoe.fi
    description: Production (backend Go)
  - url: http://localhost:8080
    description: Développement local
...
security:
  - apiKeyAuth: []
paths:
...
  /v1/webhooks: # ⚠️ cible plateforme (à construire)
    get:
      tags: [webhooks]
      operationId: listWebhookSubscriptions
      summary: Liste les abonnements webhooks de la publication
      responses:
        '200':
          description: Abonnements
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookSubscription'
    post:
      tags: [webhooks]
      operationId: createWebhookSubscription
      summary: Crée un abonnement webhook (URL + événements)
      description: >-
        Les livraisons sont signées HMAC-SHA256 (`X-Qoe-Signature`) avec un
        secret par abonnement. Retries avec backoff exponentiel et logs de
        livraison visibles dans le dashboard.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSubscriptionInput'
      responses:
        '201':
          description: Abonnement créé (secret à récupérer une seule fois)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/WebhookSubscription'
  /v1/webhooks/{id}:
    delete:
      tags: [webhooks]
      operationId: deleteWebhookSubscription
      summary: Supprime un abonnement webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Abonnement supprimé
...
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      description: >-
        Clé API `qoe_live_*` (délivrée après revue manuelle du média). Scopes
        (moindre privilège) : READ (articles/catégories), WRITE
        (créer/modifier/publier/supprimer), ANALYTICS (stats). Une clé n'a
        accès qu'aux endpoints de ses scopes — 403 sinon.
    jwtAuth:
      type: http
      scheme: bearer
      description: JWT Supabase (app mobile)
    oauth2: # ⚠️ à construire
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.qoe.fi/authorize
          tokenUrl: https://auth.qoe.fi/token
          scopes:
            articles:read: Lire les articles
            articles:write: Créer/modifier des articles
            analytics:read: Lire les statistiques
...
  schemas:
...
    WebhookSubscription: # ⚠️ cible plateforme
      type: object
      required: [id, url, events, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
    WebhookSubscriptionInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/WebhookEventType'
    WebhookEventType:
      type: string
      enum:
        - article.published
        - article.updated
        - article.deleted
        - article.scheduled
        - subscriber.created
...